1. Память делится на колличество заданных осциллограмм 2. Исправлена ошибка с размером под осциллограмму 3. Ошибка с размером одной выборки (на плате) 4. Сохранение в COMTRADE исправлено simplerate 5. При подключении к плате происходит перерасчет памяти 6. Добавлен пармаетр запуска с атрибутом: "a"

+++ Scope (Client)/ScopeSetupApp/ScopeConfig.cs
@@ -47,8 +47,8 @@ namespace ScopeSetupApp
         }
 
         //Размер всей памяти 
-        static ulong oscilAllSize = 0;
-        public static ulong OscilAllSize
+        static uint oscilAllSize = 0;
+        public static uint OscilAllSize
         {
             get { return oscilAllSize; }
             set { oscilAllSize = value; }
